fix: Disable clicking a card that has been shown
Players should not be able to click on a card that has already been shown. Before fixing this bug, a player could click on the same card as many times as they wanted, which would just toggle the card, and increase the move counter on each click.

Now, the move counter also does not increase when the player clicks on the same card.

Besides these, a change was also made to the array variable that stores the card symbols. Now, only unique symbols are stored in the array, and they are later duplicated when shuffling the cards to reset their symbols.